==> src/R/01_DBR_RBR_counting.R <==
############################################################################
############################### Script 1 ###################################
############################################################################

## Function : Summarize DNA-binding regions (DBRs) and RNA-binding regions (RBRs)
#             based on TFO-TTS pairs identified by TRIPLEX algorithm

## Input : Requires user to set a WD/path that contains the processed
#          files from the triplex quantification step.

## Output : Generates two main files. One with a summary of all DBRs and one with
#           all RBRs for all lncRNAs in a tab-separated txt file.
#           It also generates similar DBR and RBS files for each lncRNA individually.

## Libraries and helper functions
library(tidyverse)
args = commandArgs(trailingOnly=TRUE)
if(length(args)==0){
  DIR <- '/projectnb/wax-dk/alan/2022_TriplexAnalysis'
  param <- read_tsv(paste0(DIR,'/data/parameters.txt'))[1,]
}else{
  DIR <- paste0(args[1])
  param <- read_tsv(paste0(DIR,'/data/parameters.txt'))[args[2],]
}

source(paste0(DIR,"/src/helpers/functions.R"))

####################################################################################
################################# User input #######################################
####################################################################################

# Change this to the appropriate data directory where the triplex output files are located
setwd(paste0(DIR,"/output/",param$Output))

### User parameters ###
# The minimum threshold required for DNA or RNA-binding region (at least 2 TTS e.g.)
MIS_THRES <- param$MIS_THRES

# The column name in the intersected bed file with the gene chromosome number
# This signals the start of the merged triplex data columns.
GENE_CHR_COL <- param$GENE_CHR_COL

## Parameters for filtering low scoring TTS prior to grouping them into regions
SCORE = ifelse(param$SCORE == "NULL",NA,param$SCORE) #20
ERROR = ifelse(param$ERROR == "NULL",NA,param$ERROR) #0.15


####################################################################################
#################################### Script ########################################
####################################################################################

## Read-in text files generated by TRIPLEXES (implemented through the rgt-data pipeline)
file <- list.files(recursive = T,path = '00/')[grep("intersect_red.bed",list.files(recursive = T,path = '00/'))]
no_TTS_list <- NULL

# Loop through each file
for(i in 1:length(file)){
    f_tsv <-read_tsv(paste0('00/',file[i]),col_names = F)
    file_name <- unlist(strsplit(file[i],split = "/"))
    print(file[i])
    
    if(nrow(f_tsv)>=2){
      colnames(f_tsv) <- c(paste0('col_',colnames(f_tsv[,(1:(GENE_CHR_COL-1))])),
                           'Chr','start','end','lncRNA','lnc_start','lnc_end','lnc_score','lnc_error','lnc_guanineContent','lnc_DBSEnd')
      
      f_tsv <- f_tsv %>% mutate(lnc_chr = unlist(strsplit(lncRNA,split = "_"))[3])
      
      # Identify RNA-binding regions and DNA-binding regions
      f <- f_tsv %>% 
        select(gname = col_X4,
               gene_chr = Chr,
               gene_start = start,
               gene_end = end,
               lnc_chr = lnc_chr,
               start = lnc_start,
               end = lnc_end,
               score = lnc_score,
               error_rate = lnc_error,
               gc_error = lnc_guanineContent,
               length = lnc_DBSEnd,
               name=lncRNA) %>%
        mutate(min_start = min(start),
               max_end = max(end),
               gene_span=max_end-min_start) %>%
        # Arrange and calculate DNA-binding regions
        arrange(gname,
                start) %>% 
        mutate(lag_gene = lag(gname),
               lag_end = lag(end)) %>%
        replace_na(list(lag_gene = "None")) %>%
        mutate(gene_mismatch = ifelse(gname == lag_gene,T,F),
               lag_end = ifelse(!gene_mismatch,1000000,lag(end))) %>%
        group_by(gname) %>% 
        mutate(overlap = start-lag_end,
               DBR = find_domain(overlap,MIS_THRES)) %>%
        # Arrange and calculate RNA-binding regions
        arrange(gname,
                gene_start) %>%
        mutate(lag_gene = lag(gname),
               lag_gene_end = lag(gene_end)) %>%
        replace_na(list(lag_gene = "None")) %>%
        mutate(gene_mismatch = ifelse(gname == lag_gene,T,F),
               lag_gene_end = ifelse(!gene_mismatch,1000000000,lag(gene_end))) %>%
        group_by(gname) %>%
        mutate(gene_overlap = gene_start-lag_gene_end,
               RBR = find_domain(gene_overlap,MIS_THRES),
               lnc=name) %>%
        select(-name) %>%
        relocate(lnc,.before=gname)
      
      # Remove TTS that do not meet minimum criteria based on score and error rate
      if ( !is.na(SCORE) | !is.na(ERROR)){
        if(is.na(SCORE)){
          f_filter <- f[f$error_rate <= ERROR,]
        }else{
          if(is.na(ERROR)){
            f_filter <- f[f$score >= SCORE,]
          }else{
            f_filter <- f[f$score >= SCORE & f$error_rate <= ERROR,]
          }
        }
      }else{f_filter <- f}
      
      ## Summarize by unique combinations of DNA/RNA-binding regions
      f_rbr_dbr <- f_filter %>% 
        group_by(lnc,DBR,gname,RBR) %>%
        mutate(gene_start = gene_start) %>%
        summarise(RBR_DBR_avg_len = mean(length),
                  RBR_DBR_TFO_count = n(),
                  RBR_DBR_start = min(gene_start),
                  RBR_DBR_end = max(gene_end),
                  RBR_DBR_score = mean(score),
                  RBR_DBR_error_rate = mean(error_rate),
                  RBR_DBR_gc_error = mean(gc_error),
                  RBR_DBR_RBRspan = span_cal(gene_start,gene_end),
                  RBR_DBR_DBRspan = span_cal(start,end),
                  RBR_DBR_TFO_avg_overlap = sum(length)/RBR_DBR_RBRspan,
                  RBR_DBR_TFO_adj1 = RBR_DBR_RBRspan/RBR_DBR_avg_len,
                  RBR_DBR_TFO_adj2 = RBR_DBR_TFO_adj1*RBR_DBR_TFO_count) %>%
        mutate(RBR_DBR_ID = paste0(lnc,"_DBR",DBR,"_",gname,"_RBR",RBR)) %>%
        relocate(RBR_DBR_ID,.after=RBR)

      ## Summarize by RNA-binding region
      f_rbr_A <- f_filter %>% 
        group_by(lnc,gname,RBR) %>% 
        summarise(RBR_avg_len = mean(length),
                  RBR_chr = gene_chr[1],
                  RBR_start = min(gene_start),
                  RBR_end = max(gene_end),
                  RBR_span=max(gene_end)-min(gene_start)) %>%
        mutate(RBR_ID = paste0(lnc,"_",gname,"_RBR",RBR)) %>%
        relocate(RBR_ID,.before=RBR)
      
      # Summarize each DNA-binding region based on RNA-binding region table
      f_rbr_B <- f_rbr_dbr %>%
        group_by(lnc,gname,RBR) %>%
        summarise(RBR_TFO_count_total = sum(RBR_DBR_TFO_count),
                  RBR_DBR_count = n_distinct(DBR),
                  RBR_TFO_count_avg = mean(RBR_DBR_TFO_count),
                  RBR_DBR_span = sum(RBR_DBR_DBRspan),
                  RBR_TFO_abj1 = sum(RBR_DBR_TFO_adj1),
                  RBR_TFO_abj2 = sum(RBR_DBR_TFO_adj2),
                  RBR_score = mean(RBR_DBR_score),
                  RBR_error_rate = mean(RBR_DBR_error_rate),
                  RBR_gc_error = mean(RBR_DBR_gc_error),
                  BR_list = paste(RBR_DBR_ID,collapse = ",")) %>%
        mutate(RBR_ID = paste0(lnc,"_",gname,"_RBR",RBR)) %>%
        relocate(RBR_ID,.before=RBR)
      
      f_rbr_B <- re_assign(f_rbr_B,c('lnc','gname','RBR'))
      
      f_RBR <- left_join(f_rbr_A,f_rbr_B,by=c("RBR_ID"))
      
      ## Summarize by DNA-binding region
      # Summarize the coordinates for each DBR 
      f_dbr_A <- f_filter %>%
        group_by(lnc,gname,DBR) %>%
        summarize(DBR_chr = lnc_chr[1],
                  DBR_start = min(start),
                  DBR_end = max(end),
                  DBR_span = DBR_end - DBR_start) %>%
        mutate(DBR_ID = paste0(lnc,"_",gname,"_DBR",DBR)) %>%
        relocate(DBR_ID,.before=DBR)
      
      # Summarize each DNA-binding region based on RNA-binding region table
      f_dbr_B <- f_rbr_dbr %>%
        group_by(lnc,gname,DBR) %>%
        summarise(DBR_RBR_count = n_distinct(RBR),
                  DBR_TFO_count = sum(RBR_DBR_TFO_count),
                  DBR_RBR_span = sum(RBR_DBR_RBRspan),
                  DBR_TFO_abj1 = sum(RBR_DBR_TFO_adj1),
                  DBR_TFO_abj2 = sum(RBR_DBR_TFO_adj2),
                  DBR_score = mean(RBR_DBR_score),
                  DBR_error_rate = mean(RBR_DBR_error_rate),
                  DBR_gc_error = mean(RBR_DBR_gc_error),
                  BR_list = paste(RBR_DBR_ID,collapse = ",")) %>%
        mutate(DBR_ID = paste0(lnc,"_",gname,"_DBR",DBR)) %>%
        relocate(DBR_ID,.before=DBR)
      
      f_dbr_B <- re_assign(f_dbr_B,c('lnc','gname','DBR'))
      
      f_DBR <- left_join(f_dbr_A,f_dbr_B,by=c("DBR_ID"))
      
      #Save DBRs and RBRs for each individual lncRNA
      dir.create(paste0("01/",file_name[1]),recursive = T,showWarnings = F)
      write_tsv(f_rbr_dbr,file = paste0("01/",file_name[1],"/",file_name[1],"_RBRs_DBRs.txt"))
      write_tsv(f_RBR,file = paste0("01/",file_name[1],"/",file_name[1],"_RBRs.txt"))
      write_tsv(f_DBR,file = paste0("01/",file_name[1],"/",file_name[1],"_DBRs.txt"))
      
    }else{
      no_TTS_list <- c(no_TTS_list,file_name[1]) 
    }
}

meta_list <- list(paste0("# Number of records: ",length(file)),
     paste0("# Number of records processed: ",i),
     paste0("# Number of records processed: ",i),
     paste0("# Number of lncRNA with no DBR: ",length(no_TTS_list)),
     paste0("# List of lncRNA with no DBR: ",no_TTS_list),
     " ",
     "#### Parameters #####",
     paste0("# Minimum Threshold: ",MIS_THRES),
     paste0("# Chromosome ID column: ",GENE_CHR_COL),
     paste0("# Score Threshold (NA = none): ",SCORE),
     paste0("# Error Threshold (NA = none): ",ERROR))
writeLines(unlist(meta_list),'01/metadata.txt')

#####################################################
###### Script for combining DBR and RBR files #######
#####################################################

group_DBR <- list.files(path = "01/",pattern="[0-9]_DBRs.txt",recursive = T,all.files = T,full.names = T) %>%
  lapply(read_tsv) %>%
  bind_rows()
write_tsv(group_DBR,file = "01/alllncRNA_DBRs.txt")

group_RBR <- list.files(path = "01/",pattern="[0-9]_RBRs.txt",recursive = T,all.files = T,full.names = T) %>%
  lapply(read_tsv) %>%
  bind_rows()
write_tsv(group_RBR,file = "01/alllncRNA_RBRs.txt")
